changes the messaging protocol
Before, the response could be either `OK` or `Error` with message,
but now protocol more looks like json-rtp.

`result` can be of any type depending on the method being executed.

`message` has been removed and replaced with `error`, the value of which
is set if an error occurs

`GetterRecordingInProgress` method also was added.